requirements.txt
tensorflow>=2.12.0
numpy
pandas
matplotlib
scikit-learn
opencv-python
pillow
streamlit
jupyter
ipykernel
